refactor: events, store 폴더 이동
